refactor: add angle getter functions and convert PID to use radians
- Add pulse_counter_get_angle_degrees() and pulse_counter_get_angle_radians()
  to encoder module for direct angle reading
- Update pid_controller to use radians internally:
  - Change g_absolute_setpoint from int16_t to float
  - Convert dead-band and setpoint offset calculations to radians
  - Update pid_set_absolute_setpoint() and pid_get_setpoint() to accept/return float
- Update button_handler to pass radians directly to setpoint (fallen_angle + PI)
- Update lcd_controller to use pulse_counter_get_angle_degrees()
- Update all call sites to use new angle getter functions instead of raw encoder counts

+++ include/pid_controller.h
@@ -38,10 +38,10 @@ void pid_set_kd(float kd);
 
 /**
  * @brief Devuelve el punto de consigna (setpoint) actual del controlador.
- * @return El setpoint actual en cuentas del encoder.
+ * @return El setpoint actual en radianes.
  */
-int16_t pid_get_setpoint(void);
-void pid_set_absolute_setpoint(int16_t new_setpoint);
+float pid_get_setpoint(void);
+void pid_set_absolute_setpoint(float new_setpoint);
 
 /**
  * @brief Devuelve si el bucle de control del PID está actualmente habilitado.
